new Combat features, auto battles. bare functioning game

+++ Task1_RPG_Battles/Interfaces/CharacterAttributes/ICharacterAttributes.cs
@@ -10,7 +10,7 @@ public interface ICharacterAttributes
     double Attack { get; set; }
     double Defense { get; }
     double Speed { get; }
-    public int Level { get; }
+    public double Level { get; }
     public double Experience { get; }
     public double ExperienceToNextLevel { get; }
     public double Gold { get; }
