+++ Task1_RPG_Battles/Interfaces/CharacterAttributes/IMageAttributes.cs
@@ -2,8 +2,8 @@ namespace ADS_A1.Interfaces.CharacterAttributes;
 
 public interface IMageAttributes : ICharacterAttributes
 {
-    int Mana { get; set; }
-    int MaxMana { get; set; }
+    double Mana { get; set; }
+    double MaxMana { get; set; }
     // Runes are a placeholder for a future feature for new spells
-    int Runes { get; set; }
+    double Runes { get; set; }
 }
