+++ Task1_RPG_Battles/Interfaces/CharacterAttributes/IWarriorAttributes.cs
@@ -2,6 +2,6 @@ namespace ADS_A1.Interfaces.CharacterAttributes;
 
 public interface IWarriorAttributes : ICharacterAttributes
 {
-    int Rage { get; set; }
-    int MaxRage { get; set; }
+    double Rage { get; set; }
+    double MaxRage { get; set; }
 }
